extract configuration
